rename models file and remove data_access.py, this will be moved to __init__
