+++ ticketapi/datalayer/__session_wrapper.py
@@ -1,4 +1,4 @@
-from __sql_model import db, Company, Session, Ticket
+from models import db, Company, Session, Ticket
 
 class DB(object):
     def __enter__(self):
