--- ticketapi/datalayer/data_access.py
@@ -1,3 +1,0 @@
-from __session_wrapper import DB
-from __sql_model import Company, Session, Ticket
-from __stored_procedures import authenticate, update_employee, submit_ticket
